+++ tests/snapshot_tests/test_snapshots.py
@@ -1588,3 +1588,39 @@ def test_escape_to_minimize(snap_compare):
 
     # ctrl+m to maximize, escape should minimize
     assert snap_compare(TextAreaExample(), press=["ctrl+m", "escape"])
+
+
+def test_escape_to_minimize_disabled(snap_compare):
+    """Regression test for https://github.com/Textualize/textual/issues/4949"""
+
+    TEXT = """\
+    def hello(name):
+        print("hello" + name)
+
+    def goodbye(name):
+        print("goodbye" + name)
+    """
+
+    class TextAreaExample(App):
+        # Disables escape to minimize
+        ESCAPE_TO_MINIMIZE = False
+        BINDINGS = [("ctrl+m", "screen.maximize")]
+        CSS = """
+        Screen {
+            align: center middle;
+        }
+
+        #code-container {
+            width: 20;
+            height: 10;
+        }
+        """
+
+        def compose(self) -> ComposeResult:
+            with Vertical(id="code-container"):
+                text_area = TextArea.code_editor(TEXT)
+                text_area.cursor_blink = False
+                yield text_area
+
+    # ctrl+m to maximize, escape should *not* minimize
+    assert snap_compare(TextAreaExample(), press=["ctrl+m", "escape"])